requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-phonenumber-field==6.4.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
idna==3.3
install==1.3.5
phonenumbers==8.12.54
pytz==2022.1
requests==2.28.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
